feat: remove checkpoint restore commadn
